All indices combined

Function to all the indices related to biomass/ yield under different growth conditions

Usage

all_indices(Yp, Ys, Mp, Ms)

Arguments

Yp

Yield under control condition

Ys

Yield under stress condition

Mp

Mean yield of all the genotypes under control condition

Ms

Mean yiels of all the genotyps under Stress condition

Value

Indices Combined

Examples

Mp <- mean(yield_data$Yp)
Ms <- mean(yield_data$Ys)
Yp <- yield_data$Yp
Ys <- yield_data$Ys
all_indices(Yp, Ys, Mp, Ms)
#> $StressToleranceIndex
#>  [1] 0.7272727 0.5833333 0.6551724 0.6271186 0.6964286 0.6315789 0.5483871
#>  [8] 0.6379310 0.6440678 0.6315789 0.5409836 0.5833333 0.6315789 0.5079365
#> [15] 0.6440678 0.6315789 0.6440678 0.6379310 0.5409836 0.6440678 0.7272727
#> [22] 0.5483871 0.6379310 0.5833333 0.6315789 0.6440678 0.6379310 0.6315789
#> [29] 0.5483871 0.6379310 0.6440678 0.6315789 0.5409836 0.5833333 0.6315789
#> [36] 0.5079365 0.6440678 0.6315789 0.6440678 0.6379310 0.5409836 0.6440678
#> [43] 0.7272727 0.5483871 0.6379310 0.5833333 0.6315789 0.6440678 0.6379310
#> [50] 0.6315789
#> 
#> $MeanProductivityIndex
#>  [1] 4750 4750 4800 4800 4750 4650 4800 4750 4850 4650 4700 4750 4650 4750 4850
#> [16] 4650 4850 4750 4700 4850 4750 4800 4750 4750 4650 4850 4750 4650 4800 4750
#> [31] 4850 4650 4700 4750 4650 4750 4850 4650 4850 4750 4700 4850 4750 4800 4750
#> [46] 4750 4650 4850 4750 4650
#> 
#> $GeometricMeanProductivity
#>  [1] 4690.416 4582.576 4694.678 4672.259 4673.329 4529.901 4591.296 4632.494
#>  [9] 4734.976 4529.901 4486.647 4582.576 4529.901 4489.989 4734.976 4529.901
#> [17] 4734.976 4632.494 4486.647 4734.976 4690.416 4591.296 4632.494 4582.576
#> [25] 4529.901 4734.976 4632.494 4529.901 4591.296 4632.494 4734.976 4529.901
#> [33] 4486.647 4582.576 4529.901 4489.989 4734.976 4529.901 4734.976 4632.494
#> [41] 4486.647 4734.976 4690.416 4591.296 4632.494 4582.576 4529.901 4734.976
#> [49] 4632.494 4529.901
#> 
#> $Tolerance_Index
#>  [1] 1500 2500 2000 2200 1700 2100 2800 2100 2100 2100 2800 2500 2100 3100 2100
#> [16] 2100 2100 2100 2800 2100 1500 2800 2100 2500 2100 2100 2100 2100 2800 2100
#> [31] 2100 2100 2800 2500 2100 3100 2100 2100 2100 2100 2800 2100 1500 2800 2100
#> [46] 2500 2100 2100 2100 2100
#> 
#> $StressSusceptibilityIndex
#>  [1] 1.1781881 0.9449713 1.0613710 1.0159171 1.1282131 1.0231350 0.8883527
#>  [8] 1.0334329 1.0433817 1.0231350 0.8763517 0.9449713 1.0231350 0.8228101
#> [15] 1.0433817 1.0231350 1.0433817 1.0334329 0.8763517 1.0433817 1.1781881
#> [22] 0.8883527 1.0334329 0.9449713 1.0231350 1.0433817 1.0334329 1.0231350
#> [29] 0.8883527 1.0334329 1.0433817 1.0231350 0.8763517 0.9449713 1.0231350
#> [36] 0.8228101 1.0433817 1.0231350 1.0433817 1.0334329 0.8763517 1.0433817
#> [43] 1.1781881 0.8883527 1.0334329 0.9449713 1.0231350 1.0433817 1.0334329
#> [50] 1.0231350
#> 
#> $YSI
#>  [1] 0.7272727 0.5833333 0.6551724 0.6271186 0.6964286 0.6315789 0.5483871
#>  [8] 0.6379310 0.6440678 0.6315789 0.5409836 0.5833333 0.6315789 0.5079365
#> [15] 0.6440678 0.6315789 0.6440678 0.6379310 0.5409836 0.6440678 0.7272727
#> [22] 0.5483871 0.6379310 0.5833333 0.6315789 0.6440678 0.6379310 0.6315789
#> [29] 0.5483871 0.6379310 0.6440678 0.6315789 0.5409836 0.5833333 0.6315789
#> [36] 0.5079365 0.6440678 0.6315789 0.6440678 0.6379310 0.5409836 0.6440678
#> [43] 0.7272727 0.5483871 0.6379310 0.5833333 0.6315789 0.6440678 0.6379310
#> [50] 0.6315789
#> 
#> $YR_Ratio
#>  [1] 0.2727273 0.4166667 0.3448276 0.3728814 0.3035714 0.3684211 0.4516129
#>  [8] 0.3620690 0.3559322 0.3684211 0.4590164 0.4166667 0.3684211 0.4920635
#> [15] 0.3559322 0.3684211 0.3559322 0.3620690 0.4590164 0.3559322 0.2727273
#> [22] 0.4516129 0.3620690 0.4166667 0.3684211 0.3559322 0.3620690 0.3684211
#> [29] 0.4516129 0.3620690 0.3559322 0.3684211 0.4590164 0.4166667 0.3684211
#> [36] 0.4920635 0.3559322 0.3684211 0.3559322 0.3620690 0.4590164 0.3559322
#> [43] 0.2727273 0.4516129 0.3620690 0.4166667 0.3684211 0.3559322 0.3620690
#> [50] 0.3684211
#> 
#> $DRI
#>  [1] 0.5289256 0.3402778 0.4292509 0.3932778 0.4850128 0.3988920 0.3007284
#>  [8] 0.4069560 0.4148233 0.3988920 0.2926633 0.3402778 0.3988920 0.2579995
#> [15] 0.4148233 0.3988920 0.4148233 0.4069560 0.2926633 0.4148233 0.5289256
#> [22] 0.3007284 0.4069560 0.3402778 0.3988920 0.4148233 0.4069560 0.3988920
#> [29] 0.3007284 0.4069560 0.4148233 0.3988920 0.2926633 0.3402778 0.3988920
#> [36] 0.2579995 0.4148233 0.3988920 0.4148233 0.4069560 0.2926633 0.4148233
#> [43] 0.5289256 0.3007284 0.4069560 0.3402778 0.3988920 0.4148233 0.4069560
#> [50] 0.3988920
#> 
#> $HAM
#>  [1] 4631.579 4421.053 4591.667 4547.917 4597.895 4412.903 4391.667 4517.895
#>  [9] 4622.680 4412.903 4282.979 4421.053 4412.903 4244.211 4622.680 4412.903
#> [17] 4622.680 4517.895 4282.979 4622.680 4631.579 4391.667 4517.895 4421.053
#> [25] 4412.903 4622.680 4517.895 4412.903 4391.667 4517.895 4622.680 4412.903
#> [33] 4282.979 4421.053 4412.903 4244.211 4622.680 4412.903 4622.680 4517.895
#> [41] 4282.979 4622.680 4631.579 4391.667 4517.895 4421.053 4412.903 4622.680
#> [49] 4517.895 4412.903
#> 
#> $Y_Index
#>  [1] 1.1031440 0.9652510 1.0479868 1.0204082 1.0755654 0.9928296 0.9376724
#>  [8] 1.0204082 1.0479868 0.9928296 0.9100938 0.9652510 0.9928296 0.8825152
#> [15] 1.0479868 0.9928296 1.0479868 1.0204082 0.9100938 1.0479868 1.1031440
#> [22] 0.9376724 1.0204082 0.9652510 0.9928296 1.0479868 1.0204082 0.9928296
#> [29] 0.9376724 1.0204082 1.0479868 0.9928296 0.9100938 0.9652510 0.9928296
#> [36] 0.8825152 1.0479868 0.9928296 1.0479868 1.0204082 0.9100938 1.0479868
#> [43] 1.1031440 0.9376724 1.0204082 0.9652510 0.9928296 1.0479868 1.0204082
#> [50] 0.9928296
#> 
#> $yield_reduction
#>  [1] 37.50000 71.42857 52.63158 59.45946 43.58974 58.33333 82.35294 56.75676
#>  [9] 55.26316 58.33333 84.84848 71.42857 58.33333 96.87500 55.26316 58.33333
#> [17] 55.26316 56.75676 84.84848 55.26316 37.50000 82.35294 56.75676 71.42857
#> [25] 58.33333 55.26316 56.75676 58.33333 82.35294 56.75676 55.26316 58.33333
#> [33] 84.84848 71.42857 58.33333 96.87500 55.26316 58.33333 55.26316 56.75676
#> [41] 84.84848 55.26316 37.50000 82.35294 56.75676 71.42857 58.33333 55.26316
#> [49] 56.75676 58.33333
#> 
#> $R_drought_index
#>  [1] 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966
#>  [8] 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966
#> [15] 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966
#> [22] 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966
#> [29] 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966
#> [36] 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966
#> [43] 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966 0.6172966
#> [50] 0.6172966
#> 
#> $Golden_mean
#>  [1] 6.333333 3.800000 4.800000 4.363636 5.588235 4.428571 3.428571 4.523810
#>  [9] 4.619048 4.428571 3.357143 3.800000 4.428571 3.064516 4.619048 4.428571
#> [17] 4.619048 4.523810 3.357143 4.619048 6.333333 3.428571 4.523810 3.800000
#> [25] 4.428571 4.619048 4.523810 4.428571 3.428571 4.523810 4.619048 4.428571
#> [33] 3.357143 3.800000 4.428571 3.064516 4.619048 4.428571 4.619048 4.523810
#> [41] 3.357143 4.619048 6.333333 3.428571 4.523810 3.800000 4.428571 4.619048
#> [49] 4.523810 4.428571
#> 
#> $ATI
#>  [1] 4343066 7072021 5796017 6345172 4904210 5872213 7935736 6005207 6138058
#> [10] 5872213 7754857 7072021 5872213 8592130 6138058 5872213 6138058 6005207
#> [19] 7754857 6138058 4343066 7935736 6005207 7072021 5872213 6138058 6005207
#> [28] 5872213 7935736 6005207 6138058 5872213 7754857 7072021 5872213 8592130
#> [37] 6138058 5872213 6138058 6005207 7754857 6138058 4343066 7935736 6005207
#> [46] 7072021 5872213 6138058 6005207 5872213
#>