Stress susceptibility index (SSI) by Fischer and Maurer (1978)

The genotype with high SSI < 1 are more resistant to drought stress conditions

ss_index(Yp, Ys, Ms, Mp)

Arguments

Yp

Yield under control condition

Ys

Yield under Stress condition

Ms

Mean yield of all the genotypes under control condition

Mp

Mean yield of all the genotypes under stress condition

Value

SSI

References

Fischer, R. A., & Maurer, R. (1978). Drought resistance in spring wheat cultivars. I. Grain yield responses. Australian Journal of Agricultural Research, 29(5), 897-912.

Examples

ss_index(500, 350, 450, 370)
#> [1] 0.5751893